back/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
Brotli==1.1.0
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.3
decorator==5.1.1
dj-database-url==2.1.0
Django==3.2.24
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.1
exceptiongroup==1.2.0
executing==2.0.1
firebase-admin==6.4.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
ipython==7.34.0
jedi==0.19.1
jsonschema==4.17.3
jsonschema-specifications==0.1.3
matplotlib-inline==0.1.6
model-bakery==1.13.0
msgpack==1.0.5
multidict==6.0.5
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
pluggy==1.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.24.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pyrsistent==0.19.3
pytest==7.4.4
pytest-django==4.5.2
pytest-testdox==3.1.0
python-dotenv==0.21.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.8.9
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.7
wcwidth==0.2.13
whitenoise==6.5.0
yarl==1.9.4
zipp==3.15.0
